requirements.txt
google-generativeai>=0.5.2
IPython>=7.31.0
